requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
asyncpg==0.29.0